Fusing trip data manager and data processor for less fragmentation, resolving race condition and timing issues regarding data base writes and trip deltas
